Main

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2022/23.

Nota: Para a realização desta tarefa, foram importadas todas as tarefas anteriores bem como algumas bibliotecas complementares ao random, System e Gloss.

Bibliotecas utilizadas:

import System.Random -- ^ Importa a biblioteca random (para a geração de números aleatórios)
import Graphics.Gloss -- ^ Importa o Gloss (para a execução da parte gráfica do jogo)
import Graphics.Gloss.Interface.IO.Game -- ^ Importa a função "playIO" do Gloss (para a execução do jogo)
import Graphics.Gloss.Interface.Environment -- ^ Importa a função "getScreenSize" do Gloss que retorna um IO (Int,Int) com o tamanho do ecrã
import System.IO.Unsafe -- ^ Importa a função "unsafePerformIO" do System.IO.Unsafe (não é recomendado o uso desta função, mas foi utilizada para obter o tamanho do ecrã apenas para o inicio do jogo)
import System.Exit -- ^ Importa a função "exitSuccess" do System.Exit (para terminar o jogo em segurança)

Instalação das bibliotecas:

$ cabal install --lib random
$ cabal install --lib gloss

Documentation

type Score

Arguments

= Int
Tipo que define o score do jogo

type Difficulty

Arguments

= Int
Tipo que define a dificuldade do jogo

type JogadorBMP

Arguments

= Int
Tipo que define o BMP do jogador selecionado

type SubMenu

Arguments

= Int
Tipo que define o sub-menu selecionado

type Menu

Arguments

= (Bool, Int, SubMenu, Score, Score, Float, Difficulty, JogadorBMP)
Tipo que define o menu do jogo (tuplo de tipos anterioremente definidos)

type Estado

Arguments

= (Menu, Jogo)
Tipo que define o estado do jogo (conjunto do menu e do jogo)

carregarBMP

Arguments

:: String
Caminho do ficheiro BMP

-> Picture
Retorna a imagem BMP em formato Picture utilizável pelo Gloss

Função carregarBMP que carrega uma imagem BMP nas dimensões definidas pelo tamanhobmp para o Gloss.

relvabmp :: Picture

Função relvabmp que retorna o BMP da Relva em formato Picture.

aguabmp :: Picture

Função aguabmp que retorna o BMP da água do Rio em formato Picture.

estradabmp :: Picture

Função estradabmp que retorna o BMP da Estrada em formato Picture.

jogadorbmp

Arguments

:: Int
Número do BMP do jogador (jogador selecionado - 0, 1, 2 ou 3)

-> Picture
Retorna a imagem BMP do jogador em formato Picture

Função jogadorbmp que retorna o BMP do Jogador em formato Picture dado o número do jogador selecionado.

Alternativamente a função pode ser definida da seguinte forma:

Expand

jogadorbmp :: Int -> Picture
jogadorbmp n = | n > 3 = Translate 9.0 8.0 (Scale 0.7 0.7 (carregarBMP ("jogador" ++ show n ++ ".bmp"))
               | otherwise = Translate 9.0 8.0 (Scale 0.7 0.7 (carregarBMP ("jogador.bmp"))

Exemplos de utilização:

>>> jogadorbmp 0
Translate 9.0 8.0 (Scale 0.7 0.7 (carregarBMP ("jogador.bmp"))

>>> jogadorbmp 1
Translate 9.0 8.0 (Scale 0.7 0.7 (carregarBMP ("jogador1.bmp"))

>>> jogadorbmp 2
Translate 9.0 8.0 (Scale 0.7 0.7 (carregarBMP ("jogador2.bmp"))

>>> jogadorbmp 3
Translate 9.0 8.0 (Scale 0.7 0.7 (carregarBMP ("jogador3.bmp"))

seljogadorbmp

Arguments

:: Int
Número (0 ou 1) que define se o menu está selecionado (1) ou não (0)

-> Picture
Retorna a imagem BMP do menu em formato Picture

Função seljogadorbmp que retorna a Picture do Menu "Selecionar Jogador" dado o número deste ser selecionado (1) ou não (0).

Exemplos de utilização:

>>> seljogadorbmp 0
carregarBMP ("seljogador.bmp")

>>> seljogadorbmp 1
carregarBMP ("seljogador_sel.bmp")

seljogadorselectorbmp

Arguments

:: Int
Número do jogador selecionado (0, 1, 2 ou 3)

-> Picture
Retorna a imagem BMP do selector em formato Picture

Função seljogadorselectorbmp que retorna a Picture do selector do menu "Selecionar Jogador" (Número do SubMenu) dado o número do jogador selecionado.

Caso o número do jogador selecionado não esteja definido, retorna a Picture do selector do menu sem nenhum jogador selecionado (Seleciona a opção Sair do menu).

Exemplos de utilização:

>>> seljogadorselectorbmp 0
carregarBMP ("seljogadorselector1.bmp")

>>> seljogadorselectorbmp 1
carregarBMP ("seljogadorselector2.bmp")

>>> seljogadorselectorbmp 2
carregarBMP ("seljogadorselector3.bmp")

>>> seljogadorselectorbmp 3
carregarBMP ("seljogadorselector4.bmp")

>>> seljogadorselectorbmp 4
carregarBMP ("seljogadorselector.bmp")

seldificuldadebmp

Arguments

:: Int
Número (0 ou 1) que define se o menu está selecionado (1) ou não (0)

-> Picture
Retorna a imagem BMP do menu em formato Picture

Função seldificuldadebmp que retorna a Picture do Menu "Selecionar Dificuldade" dado o número deste ser selecionado (1) ou não (0).

Exemplos de utilização:

>>> seldificuldadebmp 0
carregarBMP ("seldif.bmp")

>>> seldificuldadebmp 1
carregarBMP ("seldif_sel.bmp")

seldificuldadeselectorbmp :: Int -> Picture

Função seldificuldadeselectorbmp que retorna a Picture do selector do menu "Selecionar Dificuldade" (Número do SubMenu) dado o número da dificuldade selecionada.

Caso o número da dificuldade selecionada não esteja definido, retorna a Picture do selector do menu sem nenhuma dificuldade selecionada (Seleciona a opção Sair do menu).

Exemplos de utilização:

>>> seldificuldadeselectorbmp 0
carregarBMP ("seldifselector1.bmp")

>>> seldificuldadeselectorbmp 1
carregarBMP ("seldifselector2.bmp")

>>> seldificuldadeselectorbmp 2
carregarBMP ("seldifselector3.bmp")

>>> seldificuldadeselectorbmp 3
carregarBMP ("seldifselector4.bmp")

>>> seldificuldadeselectorbmp 4
carregarBMP ("seldifselector.bmp")

arvorebmp :: Picture

Função arvorebmp que retorna a Picture da Arvore.

troncobmp :: Picture

Função troncobmp que retorna a Picture do Tronco.

carrobmp

Arguments

:: Int
Número que define a direção do carro (positivo para a direita, negativo para a esquerda)

-> Picture
Retorna a imagem BMP do carro em formato Picture

Função carrobmp que retorna a Picture do Carro dado o um número que define a direção do carro.

Caso o número seja positivo, o carro vai para a direita, caso contrário, vai para a esquerda.

Exemplos de utilização:

>>> carrobmp 1
carregarBMP ("carrod.bmp")

>>> carrobmp 2
carregarBMP ("carro1d.bmp")

>>> carrobmp (-1)
carregarBMP ("carroe.bmp")

>>> carrobmp (-2)
carregarBMP ("carro1e.bmp")

Nota: O número serve também para escolher aleatoriamente entre as imagens dos carros para a direita ou para a esquerda, utilizando a função chooseRandom do módulo System.Random.

logobmp :: Picture

Função logobmp que retorna a Picture do logo do jogo no Menu.

instrucoespainelbmp :: Picture

Função instrucoespainelbmp que retorna a Picture do painel de instruções no SubMenu instruções.

bgbmp :: Picture

Função bgbmp que retorna a Picture do fundo do jogo.

menujogar

Arguments

:: Int
Número que define se o menu Jogar está selecionado ou não (1 para selecionado, 0 para não selecionado)

-> Picture
Retorna a imagem BMP do menu Jogar em formato Picture

Função menujogar que retorna a Picture do Menu Jogar dado o número deste ser selecionado (1) ou não (0).

Exemplos de utilização:

>>> menujogar 1
carregarBMP ("jogar_sel.bmp")

>>> menujogar 0
carregarBMP ("jogar.bmp")

menusair

Arguments

:: Int
Número que define se o menu Sair está selecionado ou não (1 para selecionado, 0 para não selecionado)

-> Picture
Retorna a imagem BMP do menu Sair em formato Picture

Função menusair que retorna a Picture do Menu Sair dado o número deste ser selecionado (1) ou não (0).

Exemplos de utilização:

>>> menusair 1
carregarBMP ("sair_sel.bmp")

>>> menusair 0
carregarBMP ("sair.bmp")

menuinstrucoes

Arguments

:: Int
Número que define se o menu Instruções está selecionado ou não (1 para selecionado, 0 para não selecionado)

-> Picture
Retorna a imagem BMP do menu Instruções em formato Picture

Função menuinstrucoes que retorna a Picture do Menu Instruções dado o número deste ser selecionado (1) ou não (0).

Exemplos de utilização:

>>> menuinstrucoes 1
carregarBMP ("instrucoes_sel.bmp")

>>> menuinstrucoes 0
carregarBMP ("instrucoes.bmp")

menugameover :: Picture

Função menugameover que retorna a Picture do Menu Game Over quando o jogador perde.

tamanhobmp :: Int

Função tamanhobmp que retorna o tamanho de cada BMP do jogo.

listaLimpa :: Int -> [Obstaculo]

Função listaLimpa que gera uma lista de [Obstaculo] apenas com Nenhum, com o tamanho dado.

Exemplos de utilização:

>>> listaLimpa 3
[Nenhum, Nenhum, Nenhum]

>>> listaLimpa 5
[Nenhum, Nenhum, Nenhum, Nenhum, Nenhum]

Propriedades:

listaLimpa 0 == []

geraMapa :: IO Mapa

Função geraMapa que gera um Mapa aleatório com um tamanho aleatório entre 6 e o tamanho máximo possível de ecrã.

Nota: Esta função utiliza a função geraMapaAux, bem como a função mapaValido das tarefas anteriores para verificar que o mapa gerado é válido.

geraMapaAux

Arguments

:: Mapa
Mapa a estender

-> Int
Número de vezes que o mapa é estendido

-> IO Mapa
Retorna o mapa estendido

Função geraMapaAux que estende um Mapa aleatório, n ::Int vezes, utilizando a função estendeMapa de tarefas anteriores, gerando-se um Mapa.

Nota: Esta função utiliza um número aleatório entre 0 e 100 para a função estendeMapa como pedido no enunciado da Tarea 2.

Propriedades:

geraMapaAux mapa 0 == mapa

getSavegame

Arguments

:: IO (Int, Int, Int)
Retorna o Score, Difficulty e JogadorBMP lidos

Função getSavegame que lê os dados salvos no ficheiro "savegame.txt" e retorna-os em forma de tuplo (Score, Difficulty, JogadorBMP).

writeSavegame

Arguments

:: Int
Score a ser escrito

-> Int
Difficulty a ser escrita

-> Int
JogadorBMP a ser escrita

-> IO (Int, Int, Int)
Retorna o Score, Difficulty e JogadorBMP escritos

Função writeSavegame que escreve os Score, Difficulty e JogadorBMP dados e salva-os no ficheiro "savegame.txt" e retorna-os em forma de tuplo (Score, Difficulty, JogadorBMP).

estadoInicial

Arguments

:: Int
Top Score inicial lido

-> Int
Difficulty inicial lido

-> Int
JogadorBMP inicial lido

-> Mapa
Mapa inicial gerado

-> Estado
Retorna o Estado inicial do jogo

Função estadoInicial que retorna o Estado inicial do jogo.

desenha :: Estado -> IO Picture

Função desenhaMenu que desenha o Estado do jogo.

desenhaMenu :: Int -> Int -> [Picture]

Função desenhaMenu que desenha o menu do jogo dado o seu Estado de Menu e SubMenu.

desenhaJogador

Arguments

:: Jogador
Estado e posição do jogador

-> Int
JogadorBPM selecionado

-> Picture
Imagem do jogador

Função desenhaJogo que desenha o jogo dada a sua posição e JogadorBPM selecionado.

desenharMapa

Arguments

:: Mapa
Mapa do jogo

-> Int
Posição x utilizada para desenhar os obstaculos e linhas do mapa (incremental)

-> Int
Posição y utilizada para desenhar os obstaculos e linhas do mapa (incremental)

-> Picture
Imagem do mapa

Função desenhaLinha que desenha o mapa do jogo.

desenhaLinha

Arguments

:: Int
Comprimento da linha / Largura do mapa

-> Int
Posição x utilizada para desenhar a linha

-> Int
Posição y utilizada para desenhar a linha

-> Picture
Imagem do BMP a utilizar como fundo da linha (Relva, Rio, Estrada)

-> Picture
Imagem da linha do mapa com o comprimento dado

Função desenhaLinha que desenha uma linha do mapa do jogo.

desenhaObstaculos

Arguments

:: [Obstaculo]
Lista de obstáculos de uma linha do mapa

-> Int
Posição x utilizada para desenhar os obstaculos

-> Int
Posição y utilizada para desenhar os obstaculos

-> Int
Velocidade do Rio ou Estrada

-> Picture
Imagem dos obstáculos da linha do mapa

Função desenhaObstaculos que desenha os obstáculos de uma linha do mapa do jogo.

evento :: Event -> Estado -> IO Estado

Função evento que reage quando o utilizador carrega numa tecla.

tempo :: Float -> Estado -> IO Estado

Função tempo que atualiza o estado do jogo em função do tempo decorrido.

frameRate

Arguments

:: Int
Difficulty dada

-> Int
Framerate a retornar

Função frameRate que define o número de frames por segundo dada uma dificuldade.

janela :: Display

Função janela que define o formato, tamanho, titulo e posição, da janela do jogo.

initScreenSize

Arguments

:: (Int, Int)
Tuplo do tamanho do ecrã (largura, altura)

Função initScreenSize que retorna o tamanho do ecrã.

main :: IO ()

Função main que inicia e corre o jogo utilizando a função playIO do Gloss.