Tarefa1_2022li1g009

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Tipos

type Linha

Arguments

= (Terreno, [Obstaculo])
Tipo que representa uma linha de um mapa.

Funções principais

mapaValido :: Mapa -> Bool

A função mapaValido que verifica se um mapa é válido.

Alternativamente, a função poderia ser definida da seguinte forma:

Expand

mapaValido mapa(Mapa l m((terreno, obstaculos):t)) | l > 0 && m /= [] && (lenghtMapa l m) && (obstaculosValidos m) && (linhaValida obstaculos) && (doisRios mapa) && (troncosValidos mapa) && (carrosValidos mapa) && (mapaContiguoValido mapa) = True
                                                     | otherwise = False

Exemplos de utilização:

>>> mapaValido (Mapa 5 [(Relva, [Arvore, Nenhum, Arvore, Nenhum, Arvore]) ,(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]) ,(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore])])
True

Propriedades:

mapaValido (Mapa _ []) = False

mapaValido (Mapa 0 _) = False

lengthMapa :: Int -> [Linha] -> Bool

A função lengthMapa que verifica se o comprimento de um mapa é igualizado pelo número de obstaculos de cada linha.

Exemplos de utilização:

>>> lengthMapa 5 [(Relva, [Arvore, Nenhum, Arvore, Nenhum, Arvore]) ,(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]) ,(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore])]
True

>>> lengthMapa 5 [(Relva, [Arvore, Nenhum, Arvore, Nenhum, Arvore]) ,(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]) ,(Relva, [Arvore, Nenhum, Nenhum, Arvore])]
False

obstaculosValidos :: [Linha] -> Bool

A função obstaculosValidos que verifica se os obstáculos de uma linha são válidos.

Exemplos de utilização:

>>> obstaculosValidos [(Relva, [Arvore, Nenhum, Arvore, Nenhum, Arvore]) ,(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]) ,(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore])]
True

>>> obstaculosValidos [(Relva, [Carro, Nenhum, Arvore, Nenhum, Arvore]) ,(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]) ,(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore]) ,(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore])]
False

Propriedades:

obstaculosValidos [] = True

Utiliza a função obstaculosValidosAux para verificar se os obstáculos de uma linha são válidos.

linhaValida :: [Linha] -> Bool

A função linhaValida que verifica se todas as linhas de um mapa, relva e estradas tem um espaço livre. E se rios tem pelo menos um tronco pelo qual o jogador pode passar.

Exemplos de utilização:

>>> linhaValida [(Relva, [Arvore, Nenhum, Arvore, Nenhum, Arvore]) ,(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]) ,(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore])]
True

>>> linhaValida [(Relva, [Arvore, Arvore, Arvore, Arvore, Arvore]) ,(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]) ,(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore])]
False

Propriedades:

linhaValida [] = True

doisRios :: Mapa -> Bool

A função doisRios que verifica se um mapa tem dois rios consecutivos não tem a mesma velocidade (direção).

Exemplos de utilização:

>>> doisRios (Mapa 5 [(Rio 1, [Nenhum, Nenhum, Nenhum, Nenhum, Nenhum]) ,(Rio 1, [Nenhum, Nenhum, Nenhum, Nenhum, Nenhum])])
False

>>> doisRios (Mapa 5 [(Rio 1, [Nenhum, Nenhum, Nenhum, Nenhum, Nenhum]) ,(Rio (-1), [Nenhum, Nenhum, Nenhum, Nenhum, Nenhum])])
True

Propriedades:

doisRios (Mapa _ []) = True

doisRios (Mapa _ (x:[])) = True

troncosValidos :: Mapa -> Bool

A função troncosValidos que verifica se troncos tem, no maximo, 5 unidades de comprimento.

Exemplos de utilização:

>>> troncosValidos (Mapa 5 [(Rio 1, [Tronco, Tronco, Nenhum, Tronco, Nenhum]) ,(Rio 1, [Nenhum, Nenhum, Nenhum, Nenhum, Nenhum])])
True

>>> troncosValidos (Mapa 5 [(Rio 1, [Tronco, Tronco, Tronco, Tronco, Tronco]) ,(Rio 1, [Nenhum, Nenhum, Nenhum, Nenhum, Nenhum])])
False

Propriedades:

troncosValidos (Mapa _ []) = True

Utiliza a função troncosValidosAux para verificar se os troncos de uma linha são válidos.

carrosValidos :: Mapa -> Bool

A função carrosValidos que verifica se carros tem, no maximo, 3 unidades de comprimento.

Exemplos de utilização:

>>> carrosValidos (Mapa 5 [(Estrada 1, [Carro, Carro, Nenhum, Carro, Nenhum]) ,(Rio 1, [Nenhum, Nenhum, Nenhum, Nenhum, Nenhum])])
True

>>> carrosValidos (Mapa 5 [(Estrada 1, [Carro, Carro, Carro, Carro, Carro]) ,(Rio 1, [Nenhum, Nenhum, Nenhum, Nenhum, Nenhum])])
False

Propriedades:

carrosValidos (Mapa _ []) = True

Utiliza a função carrosValidosAux para verificar se os carros de uma linha são válidos.

mapaContiguoValido :: Mapa -> Bool

A função mapaContiguoValido que verifica que não devem existir mais do que 4 rios continuos, nem 5 estradas ou relvas.

Exemplos de utilização:

>>> mapaContiguoValido (Mapa 3 [(Rio 1, [Nenhum, Nenhum, Nenhum]), (Rio 1, [Nenhum, Nenhum, Nenhum]), (Rio 1, [Nenhum, Nenhum, Nenhum]), (Rio 1, [Nenhum, Nenhum, Nenhum]), (Rio 1, [Nenhum, Nenhum, Nenhum])])
False

>>> mapaContiguoValido (Mapa 3 [(Rio 1, [Nenhum, Nenhum, Nenhum]), (Rio 1, [Nenhum, Nenhum, Nenhum]), (Rio 1, [Nenhum, Nenhum, Nenhum]), (Rio 1, [Nenhum, Nenhum, Nenhum])])
True

Propriedades:

mapaContiguoValido (Mapa _ []) = True

mapaContiguoValido (Mapa _ (x:[])) = True

Funções auxiliares

obstaculosValidosAux :: Terreno -> [Obstaculo] -> Bool

Função obstaculosValidosAux auxiliar que verifica se um obstáculo devia estar colocado num determinado terreno.

Exemplos de utilização:

>>> obstaculosValidosAux Relva [Arvore, Nenhum, Arvore, Nenhum, Arvore]
True

>>> obstaculosValidosAux Relva [Carro, Nenhum, Arvore, Nenhum, Arvore]
False

Propriedades:

obstaculosValidosAux _ [] = True

obstaculosValidosAux Relva [Arvore, Nenhum] = True

obstaculosValidosAux (Estrada _) [Carro, Nenhum] = True

obstaculosValidosAux (Rio _) [Tronco, Nenhum] = True

troncosValidosAux :: [Obstaculo] -> Bool

Função troncosValidosAux auxiliar que verifica na linha não existem 5 troncos consecutivos.

Exemplos de utilização:

>>> troncosValidosAux [Tronco, Tronco, Nenhum, Tronco, Nenhum]
True

>>> troncosValidosAux [Tronco, Tronco, Tronco, Tronco, Tronco]
False

Propriedades:

troncosValidosAux [] = True

troncosValidosAux [Tronco, Tronco, Nenhum, Tronco, Nenhum] = True

troncosValidosAux [Tronco, Tronco, Tronco, Tronco, Tronco] = False

carrosValidosAux :: [Obstaculo] -> Bool

Função carrosValidosAux auxiliar que verifica se na linha não existem 3 carros consecutivos.

Exemplos de utilização:

>>> carrosValidosAux [Carro, Carro, Nenhum, Carro, Nenhum]
True

>>> carrosValidosAux [Carro, Carro, Carro, Carro, Carro]
False

Propriedades:

carrosValidosAux [] = True

carrosValidosAux [Carro, Carro, Nenhum] = True

carrosValidosAux [Carro, Carro, Carro] = False