Tarefa2_2022li1g009

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Nota: Para a realização desta tarefa, foi necessário utilizar a função mapaValido da Tarefa 1, que se encontra no módulo Tarefa1_2022li1g009.

Bibliotecas utilizadas:

import System.Random

Instalação das bibliotecas:

$ cabal install --lib random

Funções principais

estendeMapa

Arguments

:: Mapa
Mapa atual

-> Int
Número aleatório que será usado para gerar um mapa

-> Mapa
Mapa estendido

A função estendeMapa que estende as linhas possiveis à frente do mapa, de forma randomica.

Exemplos de utilização:

>>> estendeMapa (Mapa 3 [(Relva, [Arvore, Nenhum, Nenhum])]) 3
Mapa 3 [(Relva,[Arvore,Nenhum,Nenhum]),(Relva,[Arvore,Nenhum,Nenhum])]

Propriedades:

Utiliza a função chooseRandom para escolher um elemento aleatório de uma lista de elementos dada pela função proximosTerrenosValidos e em seguida gera os obstáculos possiveis para esse terreno utilizando a função listaObstaculos.

proximosTerrenosValidos :: Mapa -> [Terreno]

A função proximosTerrenosValidos que determina os próximos terrenos possiveis à frente do mapa.

Exemplos de utilização:

>>> proximosTerrenosValidos (Mapa 3 [(Relva, [Arvore, Nenhum, Nenhum])])
[Rio 2,Estrada 2,Relva]

Propriedades:

proximosTerrenosValidos (Mapa l []) = [Rio _, Estrada _, Relva]

Utiliza a função velocidadeRandom para gerar um número aleatório entre -3 e 3, que será usado como velocidade do novo terreno do mapa.

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função proximosObstaculosValidos que determina os próximos obstáculos possiveis à frente do mapa.

Exemplos de utilização:

>>> proximosObstaculosValidos 3 (Relva, [Arvore, Nenhum, Nenhum])
[]

>>> proximosObstaculosValidos 3 (Relva, [Arvore, Nenhum])
[Nenhum, Arvore]

Propriedades:

proximosObstaculosValidos 0 _ = []

proximosObstaculosValidos l (_, obstaculos) | l == (length obstaculos) = []

Funções auxiliares

velocidadeRandom

Arguments

:: Int
Seed para gerar um número aleatório

-> Int
Número aleatório entre -1 e 1

A função velocidadeRandom gera um número aleatório entre -1 e 1, que será usado na função proximosTerrenosValidos para gerar um mapa.

Alternativamente a função pode ser definida entre 1 e 3 da seguinte forma:

Expand

velocidadeRandom s = (mod (head (randoms (mkStdGen s))) 3) + 1

ou entre -2 e 2 da seguinte forma:

velocidadeRandom s = (mod (head (randoms (mkStdGen s))) 5) - 2

ou:

velocidadeRandom s = fst (randomR (-2, 2) (mkStdGen s))

Exemplos de utilização:

>>> velocidadeRandom 1
1

>>> velocidadeRandom 3
-1

chooseRandom

Arguments

:: Int
Seed para gerar um número aleatório

-> [a]
Lista de elementos

-> a
Elemento aleatório da lista

A função chooseRandom escolhe um elemento aleatório de uma lista de acordo com um seed.

A função pode também ser definida da seguinte forma:

Expand

chooseRandom s [] = error "Lista vazia"
chooseRandom s xs = xs !! (mod (head (randoms (mkStdGen (s)))) (length xs))

Exemplos de utilização:

>>> chooseRandom 1 [1,2,3]
1

>>> chooseRandom 2 [1,2,3]
3

listaObstaculos

Arguments

:: Int
Número de elementos da linha pretendida

-> (Terreno, [Obstaculo])
Terreno e lista de obstáculos da linha anterior

-> Int
Seed para gerar uma lista nova

-> [Obstaculo]
Lista de obstáculos nova

A função listaObstaculos que gera uma nova lista de obstáculos para uma dada linha com l (:: Int) elementos.

Exemplos de utilização:

>>> listaObstaculos 3 (Relva, [Arvore, Nenhum]) 3
[Arvore,Nenhum,Nenhum]

>>> listaObstaculos 3 (Relva, [Arvore, Nenhum]) 2
[Nenhum,Arvore,Arvore]

Propriedades:

listaObstaculos 0 (_, []) n = []

Utiliza a função chooseRandom para escolher um obstáculo aleatório de uma lista dada pela função proximosObstaculosValidos.