Tarefa3_2022li1g009

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Nota: Para a realização desta tarefa, foi necessário importar o tipo Linha da Tarefa 1, que se encontra no módulo Tarefa1_2022li1g009.

Funções principais

animaJogo :: Jogo -> Jogada -> Jogo

A função animaJogo que anima o jogo, movendo o jogador, dado um movimento, e os obstáculos.

Exemplos de utilização:

>>> animaJogo (Jogo (Jogador (1,1)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 1, [Carro, Nenhum, Nenhum])])) (Move Cima)
Jogo (Jogador (1,2)) (Mapa 3 [(Rio 1,[Nenhum,Tronco,Nenhum]),(Estrada 1,[Carro,Nenhum,Nenhum])])

>>> animaJogo (Jogo (Jogador (1,1)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 2, [Carro, Nenhum, Nenhum])])) (Parado)
Jogo (Jogador (1,1)) (Mapa 3 [(Rio 1,[Nenhum,Tronco,Nenhum]),(Estrada 2,[Nenhum,Nenhum,Carro])])

animaJogador

Arguments

:: Jogo
Jogo inicial (estado anterior ao movimento do jogador)

-> Jogada
Jogada a ser realizada

-> Jogador
Jogador após a jogada

A função animaJogador que move o jogador segundo a jogada.

Exemplos de utilização:

>>> animaJogador (Jogo (Jogador (1,1)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 1, [Carro, Nenhum, Nenhum])])) (Move Cima)
Jogador (1,2)

>>> animaJogador (Jogo (Jogador (1,1)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 2, [Carro, Nenhum, Nenhum])])) (Parado)
Jogador (1,1)

>>> animaJogador (Jogo (Jogador (0,0)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 2, [Carro, Nenhum, Nenhum])])) (Parado)
Jogador (1,0)

Propriedades:

Caso o jogador não se mova, a função deve devolver as mesmas coordenadas do jogador, ou caso esteja parado sobre um tronco move-se segundo este utilizando a função paradoEmCimaTronco.

animaEstradaRio :: [Linha] -> [Linha]

A função animaEstradaRio que anima a estrada e o rio, movendo os obstáculos.

Função que numa estrada ou rio com velocidade v os obstáculos devem mover-se |v| unidades na direção determinada.

Exemplos de utilização:

>>> animaEstradaRio [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 1, [Carro, Nenhum, Nenhum])]
[(Rio 1,[Nenhum,Tronco,Nenhum]),(Estrada 1,[Nenhum,Carro,Nenhum])]

Propriedades:

animaEstradaRio [] = []

Funções auxiliares

rotateListRight

Arguments

:: Int
Número de vezes que a lista deve ser rotacionada n

-> [a]
Lista de elementos

-> [[a]]
Lista de listas de elementos originada da rotação da lista original por cada passo n vezes

A função rotateListRight auxiliar que roda todos os obstáculos de uma lista n vezes para a direita.

Exemplos de utilização:

>>> rotateListRight 1 [1, 2, 3]
[[1,2,3],[3,1,2]]

>>> rotateListRight 2 [1, 2, 3]
[[1,2,3],[3,1,2],[2,3,1]]

Propriedades:

rotateListRight 0 [l] = [l]

rotateListLeft

Arguments

:: Int
Número de vezes que a lista deve ser rotacionada n

-> [a]
Lista de elementos

-> [[a]]
Lista de listas de elementos originada da rotação da lista original por cada passo n vezes

A função rotateListLeft auxiliar que roda todos os obstáculos de uma lista n vezes para a esquerda.

Exemplos de utilização:

>>> rotateListLeft 1 [1, 2, 3]
[[1,2,3],[2,3,1]]

>>> rotateListLeft 2 [1, 2, 3]
[[1,2,3],[2,3,1],[3,1,2]]

Propriedades:

rotateListLeft 0 [l] = [l]

moveObstaculos :: Int -> [Obstaculo] -> [Obstaculo]

A função moveObstaculos que move os obstáculos para a direita n vezes ou para a esquerda n vezes.

Exemplos de utilização:

>>> moveObstaculos 1 [Carro, Nenhum, Nenhum]
[Nenhum,Carro,Nenhum]

>>> moveObstaculos 2 [Carro, Nenhum, Nenhum]
[Nenhum,Nenhum,Carro]

Propriedades:

moveObstaculos _ [] = []

moveObstaculos 0 [l] = [l]

naoMoveForaMapa :: Jogador -> Mapa -> Bool

A função naoMoveForaMapa que verifica se o jogador não se move para uma posição fora do mapa.

Retorna True se o jogador não se move para fora do mapa, False caso contrário.

Exemplos de utilização:

>>> naoMoveForaMapa (Jogador (1,1)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 2, [Carro, Nenhum, Nenhum])])
True

>>> naoMoveForaMapa (Jogador (1,2)) (Mapa 2 [(Rio 1, [Tronco, Nenhum]), (Estrada 2, [Carro, Nenhum])])
False

naoMoveCimaArvore :: Jogador -> Mapa -> Bool

Função auxiliar naoMoveCimaArvore que verifica se o jogador não se move para cima de uma árvore.

Exemplos de utilização:

>>> naoMoveCimaArvore (Jogador (1,1)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 1, [Carro, Nenhum, Nenhum])])
True

>>> naoMoveCimaArvore (Jogador (1,1)) (Mapa 3 [(Relva, [Arvore, Nenhum, Nenhum]), (Relva, [Nenhum, Arvore, Nenhum])]) -- False
False

paradoEmCimaTronco :: Jogador -> Mapa -> Jogador

A função paradoEmCimaTronco que retorna a posição de um jogador que está parado em cima de um tronco em movimento e move-o. Isto é, caso o jogador esteja parado em cima de um tronco, a função deve devolver as mesmas coordenadas do jogador, ou caso esteja parado sobre um tronco move-se segundo este.

Exemplos de utilização:

>>> paradoEmCimaTronco (Jogador (0,0)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 2, [Carro, Nenhum, Nenhum])])
Jogador (1,0)

>>> paradoEmCimaTronco (Jogador (1,1)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 2, [Carro, Nenhum, Nenhum])])
Jogador (1,1)

paradoEmCimaTroncoAux :: Jogador -> Mapa -> Bool

A função paradoEmCimaTroncoAux auxiliar que verifica se um jogador está parado em cima de um tronco.

Retorna True se o jogador está parado em cima de um tronco, False caso contrário.

Exemplos de utilização:

>>> paradoEmCimaTroncoAux (Jogador (0,0)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 2, [Carro, Nenhum, Nenhum])])
True

>>> paradoEmCimaTroncoAux (Jogador (1,1)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 2, [Carro, Nenhum, Nenhum])])
False

Propriedades:

paradoEmCimaTroncoAux (Jogador (x,y)) (Mapa l []) = False

velocidadeTronco :: Jogador -> Mapa -> Int

A função velocidadeTronco auxiliar que devolve a direção do tronco em que o jogador está parado.

Exemplos de utilização:

>>> velocidadeTronco (Jogador (0,0)) (Mapa 3 [(Rio 1, [Tronco, Nenhum, Nenhum]), (Estrada 2, [Carro, Nenhum, Nenhum])])
1

Propriedades:

velocidadeTronco (Jogador (x,y)) (Mapa l []) = 0

Caso o jogador não esteja parado em cima de um tronco, a função deve devolver 0.