Tarefa5_2022li1g009

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2022/23.

Documentation

deslizaJogo :: Int -> Jogo -> Jogo

Função deslizaJogo que faz com que a última linha do mapa desapareça, ao mesmo tempo que uma nova linha no topo do mapa seja criada (utilizando a função "estendeMapa" da Tarefa 2).

Exemplos de utilização:

>>> deslizaJogo 1 (Jogo (Jogador (1,0)) (Mapa 3 [(Relva, [Nenhum, Nenhum, Arvore]),(Rio 1, [Nenhum, Tronco, Nenhum]), (Estrada 1, [Carro, Nenhum, Nenhum])]))
Jogo (Jogador (1,-1)) (Mapa 3 [(Rio 1,[Nenhum,Tronco,Nenhum]),(Estrada 1,[Carro,Nenhum,Nenhum]),(Estrada (-1),[Carro,Nenhum,Carro])])

>>> deslizaJogo 2 (Jogo (Jogador (1,1)) (Mapa 3 [(Rio 1,[Nenhum,Tronco,Nenhum]),(Estrada 1,[Carro,Nenhum,Nenhum]),(Estrada 2,[Carro,Nenhum,Carro])]))
Jogo (Jogador (1,0)) (Mapa 3 [(Estrada 1,[Carro,Nenhum,Nenhum]),(Estrada 2,[Carro,Nenhum,Carro]),(Rio (-1),[Nenhum,Tronco,Tronco])])

animaDeslizaJogo :: Int -> Jogo -> Int -> Jogo

Função animaDeslizaJogo que desliza o jogo se o jogador estiver à frente de metade do mapa linha do mapa _n_ vezes.

Exemplos de utilização:

>>> animaDeslizaJogo 1 (Jogo (Jogador (1,0)) (Mapa 3 [(Relva, [Nenhum, Nenhum, Arvore]),(Rio 1, [Nenhum, Tronco, Nenhum]), (Estrada 1, [Carro, Nenhum, Nenhum])])) 1
Jogo (Jogador (1,0)) (Mapa 3 [(Relva,[Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Tronco,Nenhum]),(Estrada 1,[Carro,Nenhum,Nenhum])])

>>> animaDeslizaJogo 1 (Jogo (Jogador (1,1)) (Mapa 3 [(Rio 1,[Nenhum,Tronco,Nenhum]),(Estrada 1,[Carro,Nenhum,Nenhum]),(Estrada 2,[Carro,Nenhum,Carro])])) 3
Jogo (Jogador (1,1)) (Mapa 3 [(Rio 1,[Nenhum,Tronco,Nenhum]),(Estrada 1,[Carro,Nenhum,Nenhum]),(Estrada 2,[Carro,Nenhum,Carro])])